"finalize the project"

+++ DSP_Project/CM7/Core/Inc/DSP_Polynom.h
@@ -17,8 +17,8 @@ DSP_Return_ten DSP_PolynomInit(DSP_Polynom_tst* Polynom_st);
 DSP_Return_ten DSP_PolynomCreate(char* PolynomChar , DSP_Polynom_tst* Polynom_st);
 DSP_Return_ten DSP_SumPolynom(DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst Polynom2_st, DSP_Polynom_tst* PolynomSum_st);
 DSP_Return_ten DSP_MulPolynom(DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst Polynom2_st, DSP_Polynom_tst* PolynomMul_st);
-DSP_Return_ten DSP_MulScalarPolynom(float Scalar, DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst* PolynomMul_st);
-DSP_Return_ten DSP_DevideScalarPolynom(float Scalar, DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst* PolynomMul_st);
+DSP_Return_ten DSP_MulScalarPolynom(double Scalar, DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst* PolynomMul_st);
+DSP_Return_ten DSP_DevideScalarPolynom(double Scalar, DSP_Polynom_tst Polynom1_st, DSP_Polynom_tst* PolynomMul_st);
 DSP_Return_ten DSP_PowPolynom(DSP_Polynom_tst Polynom_st, uint8_t pow , DSP_Polynom_tst* PolynomMul_st);
 DSP_Return_ten DSP_CpyPolynom(DSP_Polynom_tst* Polynom1_st , DSP_Polynom_tst Polynom2_st);
